fix delete category

+++ back-end/controllers/categoriesControlleurs.js
@@ -36,18 +36,24 @@ exports.udapteCategory = (req, res) => {
 }
 
 exports.deleteCategory = (req, res) => {
+ 
     db.Post.findAll({ where: { categoryId: req.params.id } })
         .then((result) => {
             result.forEach(element => {
                 db.Comments.destroy({ where: { postId: element.id } })
                     .then(() => {
                         db.Post.destroy({ where: { categoryId: req.params.id } })
-                        .then(() => {
-                            db.Category.destroy({ where: { id: req.params.id } })
-                                .then(() => res.status(200).json({ message: 'categorie supprimé !' }))
-                                .catch(error => res.status(400).json({ error }));
-                        })
+                            .then(() => {
+                                db.Category.destroy({ where: { id: req.params.id } })
+                                    .then(() => res.status(200).json({ message: 'categorie supprimé !' }))
+                                    .catch(error => res.status(400).json({ error }));
+                            })
                     });
             })
+            if(result.length === 0) {
+                db.Category.destroy({ where: { id: req.params.id } })
+                    .then(() => res.status(200).json({ message: 'categorie supprimé !' }))
+                    .catch(error => res.status(400).json({ error }));
+            }
         })
 }
